requirements.txt
mss>=9.0.0
pyautogui>=0.9.54
pytesseract>=0.3.9
requests>=2.28.0
urllib3>=1.26.0
pyobjc-framework-Quartz>=9.0
pynput>=1.8.0
numpy>=1.23.0
opencv-python>=4.6.0
Pillow>=9.3.0
